+++ billing/Working/Prod/Dockerfile
@@ -3,7 +3,6 @@ RUN apt-get update
 RUN apt-get install -y python3-pip python3-dev curl
 COPY . /billingserver
 WORKDIR /billingserver
-RUN python3 changeport.py ${DYNAMIC_PORT}
 RUN pip3 install -r requirements.txt
 EXPOSE 8081
 ENTRYPOINT python3 billing.py
